Started the CRUD Operation
Main 2 functions in Password Manager. One is the main window that will have the create and read function for now.
Second is the Login Screen. We need to remember this password only and rest all will be saved by the db.

PasswordManager_Create has the create functionality. Will add more in the Readme.